chore: create UIButton extension
